Added tests for the rest API endpoints.
The following endpoints are covered.
  1. /rest/job
  2. /rest/jobexec
  3. /rest/flowexec
  4. /rest/search
  5. /rest/compare
  6. /rest/flowgraphdata
  7. /rest/jobgraphdata

The tests are run against a test server which deploys a fake application connecting to an in-memory H2 DB.
The Global class is overridden in the tests so that we don't have to go through the regular application
start flow. Assertions are made on the json response from each of the above listed endpoints.

Results from 'play test' as below.

$>play test
...
...
[info] rest.RestAPITest
[info] + restFlowExecResult
[info] + testrestJobExecResult
[info] + testrestSearch
[info] + testrestJobGraphData
[info] + testrestAppResult
[info] + testrestFlowGraphData
[info] + testrestSearchWithUsernameAndJobType
[info] + testrestCompare
[info]
[info]
[info] Total for test rest.RestAPITest
[info] Finished in 0.001 seconds
[info] 8 tests, 0 failures, 0 errors
...
...
[info] Passed: Total 141, Failed 0, Errors 0, Passed 141
[success] Total time: 5 s, completed Apr 21, 2016 3:59:15 PM