Include S3 counters in heuristics
The Tez heuristics assume that all of the S3-related counters are
prefixed with either "S3A_" or "S3N_". In our case, it's simply "S3_",
so our Mapper heuristics all show 0 bytes read per task.

This updates the heuristics to include the "S3_*" counters (the
MapReduce heuristics already do this).